Move article count into shared folder
